+++ components/home/landing-data.ts
@@ -1,0 +1,82 @@
+export const heroStats = [
+  { title: "無料掲載サービス数", value: "300", unit: "件以上" },
+  { title: "掲載ジャンル数", value: "30", unit: "カテゴリ以上" },
+  { title: "口コミ評価", value: "5", unit: "/5 点" },
+];
+
+export const serviceCategories = ["各種保険", "不動産", "エンタメ", "転職"];
+
+export const meritItems = [
+  {
+    number: "01",
+    title: "幅広いジャンルを網羅",
+    description:
+      "保険・不動産・転職・エンタメ・生活サービスなど、暮らしに関わる情報を総合的にカバー。",
+    image: "/assets/images/graph-icon.webp",
+  },
+  {
+    number: "02",
+    title: "比較・検討が一目でわかる",
+    description:
+      "各種サービス情報をまとめて掲載。自分に合った選択肢がスムーズに見つかります。",
+    image: "/assets/images/insight-icon.webp",
+  },
+  {
+    number: "03",
+    title: "無料で使える安心設計",
+    description:
+      "多くのサービスが無料で利用可能。初めての方でも安心して活用できます。",
+    image: "/assets/images/free-icon.webp",
+  },
+];
+
+export const achievementImages = [
+  { image: "/assets/images/掲載ジャンル数.webp", alt: "掲載ジャンル数20カテゴリー以上" },
+  { image: "/assets/images/無料掲載サービス数.webp", alt: "無料掲載サービス数500件以上" },
+];
+
+export const testimonials = [
+  {
+    age: "20代",
+    gender: "女性",
+    comment: "保険の見直しで年間5万円も節約できました！比較がとても簡単で助かりました。",
+    rating: 5,
+    avatar: "/assets/images/woman-icon.webp",
+  },
+  {
+    age: "30代",
+    gender: "男性",
+    comment: "転職サービスの情報が充実していて、自分に合った求人が見つかりました。サポートも丁寧でした。",
+    rating: 5,
+    avatar: "/assets/images/man-icon.webp",
+  },
+  {
+    age: "40代",
+    gender: "女性",
+    comment: "不動産査定がこんなに簡単にできるとは思いませんでした。対応も迅速で信頼できました。",
+    rating: 5,
+    avatar: "/assets/images/old-woman-icon.webp",
+  },
+  {
+    age: "50代",
+    gender: "男性",
+    comment: "エンタメサブスクの比較ができて、無駄な契約を整理できました。家計に優しいサービスです。",
+    rating: 5,
+    avatar: "/assets/images/old-man-icon.webp",
+  },
+];
+
+export const faqItems = [
+  {
+    question: "掲載するのに費用はかかりますか？",
+    answer: "掲載は完全無料です。豊富なジャンルから、あなたにぴったりのサービスを見つけましょう。",
+  },
+  {
+    question: "どんなサービスが掲載されていますか？",
+    answer: "保険・不動産・転職・エンタメなど、暮らしに役立つサービスを幅広く掲載しています。",
+  },
+  {
+    question: "スマートフォンでも利用できますか？",
+    answer: "はい、スマートフォン・タブレット・PC のいずれでも快適にご利用いただけます。",
+  },
+];
