+++ components/home/landing-page.tsx
@@ -1,119 +1,21 @@
 "use client";
 
-import { useEffect, useRef, useState } from "react";
-import Image from "next/image";
-import Link from "next/link";
 import { useSession } from "next-auth/react";
-import { useScrollReveal } from "@/hooks/use-scroll-reveal";
-import { cn } from "@/lib/utils";
-import { extractExcerpt } from "@/lib/blog-utils";
-import type { BlogResponse, CategoryResponse } from "@/types/microcms";
-
-const heroStats = [
-  {
-    title: "無料掲載サービス数",
-    value: "300",
-    unit: "件以上",
-  },
-  {
-    title: "掲載ジャンル数",
-    value: "30",
-    unit: "カテゴリ以上",
-  },
-  {
-    title: "口コミ評価",
-    value: "5",
-    unit: "/5 点",
-  },
-];
-
-const serviceCategories = ["各種保険", "不動産", "エンタメ", "転職"];
-
-const meritItems = [
-  {
-    number: "01",
-    title: "幅広いジャンルを網羅",
-    description:
-      "保険・不動産・転職・エンタメ・生活サービスなど、暮らしに関わる情報を総合的にカバー。",
-    image: "/assets/images/graph-icon.webp",
-  },
-  {
-    number: "02",
-    title: "比較・検討が一目でわかる",
-    description:
-      "各種サービス情報をまとめて掲載。自分に合った選択肢がスムーズに見つかります。",
-    image: "/assets/images/insight-icon.webp",
-  },
-  {
-    number: "03",
-    title: "無料で使える安心設計",
-    description:
-      "多くのサービスが無料で利用可能。初めての方でも安心して活用できます。",
-    image: "/assets/images/free-icon.webp",
-  },
-];
-
-const achievementImages = [
-  {
-    image: "/assets/images/掲載ジャンル数.webp",
-    alt: "掲載ジャンル数20カテゴリー以上",
-  },
-  {
-    image: "/assets/images/無料掲載サービス数.webp",
-    alt: "無料掲載サービス数500件以上",
-  },
-];
-
-type CtaResolver = (defaultLabel: string) => { href: string; label: string };
-
-const testimonials = [
-  {
-    age: "20代",
-    gender: "女性",
-    comment: "保険の見直しで年間5万円も節約できました！比較がとても簡単で助かりました。",
-    rating: 5,
-    avatar: "/assets/images/woman-icon.webp",
-  },
-  {
-    age: "30代",
-    gender: "男性",
-    comment: "転職サービスの情報が充実していて、自分に合った求人が見つかりました。サポートも丁寧でした。",
-    rating: 5,
-    avatar: "/assets/images/man-icon.webp",
-  },
-  {
-    age: "40代",
-    gender: "女性",
-    comment: "不動産査定がこんなに簡単にできるとは思いませんでした。対応も迅速で信頼できました。",
-    rating: 5,
-    avatar: "/assets/images/old-woman-icon.webp",
-  },
-  {
-    age: "50代",
-    gender: "男性",
-    comment: "エンタメサブスクの比較ができて、無駄な契約を整理できました。家計に優しいサービスです。",
-    rating: 5,
-    avatar: "/assets/images/old-man-icon.webp",
-  },
-];
-
-const faqItems = [
-  {
-    question: "掲載するのに費用はかかりますか？",
-    answer:
-      "掲載は完全無料です。豊富なジャンルから、あなたにぴったりのサービスを見つけましょう。",
-  },
-  {
-    question: "どんなサービスが掲載されていますか？",
-    answer:
-      "保険・不動産・転職・エンタメなど、暮らしに役立つサービスを幅広く掲載しています。",
-  },
-  {
-    question: "スマートフォンでも利用できますか？",
-    answer:
-      "はい、スマートフォン・タブレット・PC のいずれでも快適にご利用いただけます。",
-  },
-];
+import type { CategoryResponse } from "@/types/microcms";
+import {
+  AchievementSection,
+  BottomCtaSection,
+  FaqSection,
+  HeroSection,
+  HighlightSection,
+  IntroductionBoxSection,
+  LatestBlogsSection,
+  MeritSection,
+  ProblemSection,
+  ServiceSection,
+  TestimonialsSection,
+} from "./sections";
+import { CtaResolver } from "./types";
 
 interface LandingPageProps {
   categories: CategoryResponse[];
@@ -142,994 +44,5 @@ export function LandingPage({ categories }: LandingPageProps) {
       <FaqSection />
       <BottomCtaSection resolveCta={resolveCta} />
     </main>
-  );
-}
-
-function LatestBlogsSection() {
-  const [blogs, setBlogs] = useState<BlogResponse[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-
-  useEffect(() => {
-    const controller = new AbortController();
-    let isActive = true;
-    const LIMIT = 3;
-
-    async function fetchLatestBlogs() {
-      try {
-        setIsLoading(true);
-        const response = await fetch(`/blogs?limit=${LIMIT}`, {
-          cache: "no-store",
-          signal: controller.signal,
-        });
-
-        if (!response.ok) {
-          throw new Error(`Failed to fetch blogs: ${response.status}`);
-        }
-
-        const data = await response.json();
-
-        if (!isActive) return;
-        setBlogs(Array.isArray(data.contents) ? data.contents : []);
-        setErrorMessage(null);
-      } catch (error) {
-        if (!isActive || (error instanceof DOMException && error.name === "AbortError")) {
-          return;
-        }
-        console.error("[LandingPage] 最新ブログの取得に失敗しました", error);
-        setErrorMessage("現在ブログを取得できません。時間をおいて再度お試しください。");
-      } finally {
-        if (isActive) {
-          setIsLoading(false);
-        }
-      }
-    }
-
-    fetchLatestBlogs();
-
-    return () => {
-      isActive = false;
-      controller.abort();
-    };
-  }, []);
-
-  const formatDate = (dateString?: string) => {
-    if (!dateString) return "日付未設定";
-    try {
-      return new Intl.DateTimeFormat("ja-JP", {
-        year: "numeric",
-        month: "short",
-        day: "numeric",
-      }).format(new Date(dateString));
-    } catch {
-      return "日付未設定";
-    }
-  };
-
-  return (
-    <section className="relative -mt-16 z-20 pb-16 md:-mt-20 md:pb-24">
-      <div className="mx-auto max-w-[1100px] rounded-[36px] bg-white/80 px-6 py-12 shadow-[0_25px_65px_rgba(15,23,42,0.08)] backdrop-blur-lg lg:px-12">
-        <div className="flex flex-col gap-6 md:flex-row md:items-end md:justify-between">
-          <div className="space-y-3">
-            <p className="text-xs font-semibold uppercase tracking-[0.35em] text-win2-primary-orage">
-              Latest Blog
-            </p>
-            <h2 className="text-3xl font-bold text-slate-900 md:text-4xl">最新のブログインサイト</h2>
-            <p className="text-sm text-slate-600 md:text-base">
-              WIN×Ⅱ編集部が厳選した最新の記事から、暮らしに役立つヒントを素早くキャッチしましょう。
-            </p>
-          </div>
-          <Link
-            href="/blog"
-            className="inline-flex items-center justify-center rounded-full border border-win2-primary-orage/40 bg-white px-6 py-3 text-sm font-semibold text-win2-primary-orage shadow-[0_5px_20px_rgba(244,138,60,0.25)] transition hover:-translate-y-0.5 hover:bg-win2-primary-orage hover:text-white"
-          >
-            すべての記事を見る
-          </Link>
-        </div>
-
-        <div className="mt-10">
-          {isLoading ? (
-            <div className="grid gap-6 md:grid-cols-3">
-              {Array.from({ length: 3 }).map((_, index) => (
-                // eslint-disable-next-line react/no-array-index-key
-                <div key={index} className="h-56 animate-pulse rounded-3xl bg-slate-200/70" />
-              ))}
-            </div>
-          ) : errorMessage ? (
-            <div className="rounded-3xl border border-red-100 bg-red-50/80 p-6 text-sm text-red-700">
-              {errorMessage}
-            </div>
-          ) : blogs.length === 0 ? (
-            <div className="rounded-3xl border border-dashed border-slate-200 bg-white/70 p-6 text-center text-sm text-slate-500">
-              まだ公開中のブログ記事がありません。
-            </div>
-          ) : (
-            <div className="flex flex-col gap-6">
-              {blogs.map((blog) => {
-                const thumbnailUrl = blog.thumbnail?.url ?? "/assets/images/blog-placeholder.webp";
-                return (
-                  <Link
-                    key={blog.id}
-                    href={`/blog/${blog.id}`}
-                    className="group flex flex-col gap-5 rounded-3xl border border-white/70 bg-white/95 p-5 shadow-md shadow-slate-200/40 transition hover:-translate-y-1 hover:shadow-xl focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-win2-primary-orage/40 md:flex-row md:p-6"
-                  >
-                    <div className="relative h-48 w-full overflow-hidden rounded-2xl bg-slate-100 md:h-40 md:w-64">
-                      <Image
-                        src={thumbnailUrl}
-                        alt={blog.title}
-                        fill
-                        sizes="(min-width: 1024px) 320px, 100vw"
-                        className="object-cover transition duration-500 group-hover:scale-105"
-                      />
-                    </div>
-
-                    <div className="flex flex-1 flex-col gap-4">
-                      <div className="flex flex-wrap gap-2 text-xs font-semibold text-win2-primary-orage">
-                        <span>{blog.category?.name ?? "コラム"}</span>
-                        <span className="text-slate-400">• {formatDate(blog.publishedAt || blog.createdAt)}</span>
-                      </div>
-                      <h3 className="text-xl font-bold leading-snug text-slate-900 transition group-hover:text-win2-primary-orage">
-                        {blog.title}
-                      </h3>
-                      <p className="text-sm leading-relaxed text-slate-600 line-clamp-3">
-                        {extractExcerpt(blog.content ?? "", 140)}
-                      </p>
-                      <span className="inline-flex items-center text-sm font-semibold text-win2-primary-orage transition group-hover:gap-2">
-                        記事を読む
-                        <svg
-                          className="ml-1 h-4 w-4 transition group-hover:translate-x-1"
-                          viewBox="0 0 24 24"
-                          fill="none"
-                          xmlns="http://www.w3.org/2000/svg"
-                          aria-hidden="true"
-                        >
-                          <path
-                            d="M5 12h14M13 6l6 6-6 6"
-                            stroke="currentColor"
-                            strokeWidth="1.6"
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                          />
-                        </svg>
-                      </span>
-                    </div>
-                  </Link>
-                );
-              })}
-            </div>
-          )}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function HeroSection({ resolveCta }: { resolveCta: CtaResolver }) {
-  const { ref } = useScrollReveal<HTMLDivElement>();
-  const [isMounted, setIsMounted] = useState(false);
-  const headlineRef = useRef<HTMLHeadingElement | null>(null);
-  const imageRef = useRef<HTMLDivElement | null>(null);
-  const statsRef = useRef<HTMLDivElement | null>(null);
-  const ctaRef = useRef<HTMLDivElement | null>(null);
-  const [playKey, setPlayKey] = useState(0);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setIsMounted(true);
-    }, 80);
-    return () => clearTimeout(timer);
-  }, []);
-
-  useEffect(() => {
-    if (!isMounted) return;
-    setPlayKey((prev) => prev + 1);
-  }, [isMounted]);
-
-  return (
-    <section className="relative -mb-32 min-h-[540px] overflow-visible bg-gradient-to-b from-win2-surface-cream-320 via-white to-win2-surface-cream-150 pb-32 md:-mb-40 md:min-h-[640px] md:pb-44">
-      <div className="absolute inset-0 overflow-hidden">
-        <div className="relative h-full w-full min-h-full">
-          <Image
-            src="/assets/images/office-super-blur.webp"
-            alt="オフィス背景"
-            fill
-            sizes="100vw"
-            className={cn(
-              "origin-top opacity-0 blur-[14px] object-cover",
-              isMounted && playKey ? "animate-hero-background" : "animate-none"
-            )}
-            style={{ objectFit: "cover", objectPosition: "center top" }}
-            quality={75}
-            priority
-            loading="eager"
-          />
-        </div>
-        <div
-          className={cn(
-            "absolute inset-0 bg-gradient-to-b from-white/0 via-white/40 to-white/85 opacity-0",
-            isMounted && playKey ? "animate-hero-overlay" : "animate-none"
-          )}
-        />
-      </div>
-      <div
-        ref={ref}
-        className="relative z-10 mx-auto flex max-w-[1100px] flex-col gap-12 px-6 pt-12 md:flex-row md:items-center lg:px-8"
-      >
-        <div className="relative z-40 max-w-xl space-y-8">
-          <div className="space-y-3">
-            <p
-              className={cn(
-                "text-sm font-semibold tracking-[0.35em] text-win2-accent-gold opacity-0",
-                isMounted && playKey ? "animate-hero-badge" : "animate-none"
-              )}
-            >
-              WIN×Ⅱ
-            </p>
-            <h1
-              ref={headlineRef}
-              className={cn(
-                "text-4xl font-bold leading-tight text-win2-neutral-950 opacity-0 md:text-5xl",
-                isMounted && playKey ? "animate-hero-headline" : "animate-none"
-              )}
-            >
-              暮らしを
-              <span className="text-win2-primary-orage whitespace-nowrap">もっとお得に</span>
-              <br />
-              <span className="text-win2-primary-orage whitespace-nowrap">もっとスマートに。</span>
-            </h1>
-            <p
-              className={cn(
-                "text-[15px] leading-relaxed text-slate-700 opacity-0",
-                isMounted && playKey ? "animate-hero-subtext" : "animate-none"
-              )}
-            >
-              WIN×Ⅱは、暮らしに役立つサービスをワンストップで選べるアフィリエイトブログプラットフォームです。保険・不動産・転職・エンタメなど豊富なジャンルから、あなたにぴったりのサービスを見つけましょう。
-            </p>
-          </div>
-          <div
-            ref={statsRef}
-            className={cn(
-              "flex flex-wrap gap-4 opacity-0",
-              isMounted && playKey ? "animate-hero-stats" : "animate-none"
-            )}
-          >
-            {heroStats.map((stat) => (
-              <div
-                key={stat.title}
-                className="flex min-w-[160px] flex-1 flex-col items-center rounded-[28px] bg-white/90 px-5 py-4 text-center shadow-[0_10px_22px_rgba(240,130,90,0.18)] backdrop-blur"
-              >
-                <p className="text-xs font-semibold text-win2-primary-orage">{stat.title}</p>
-                <p className="mt-2 text-2xl font-bold text-win2-neutral-950">
-                  {stat.value}
-                  <span className="ml-1 text-base font-semibold text-slate-600 whitespace-nowrap">{stat.unit}</span>
-                </p>
-              </div>
-            ))}
-          </div>
-          <div
-            ref={ctaRef}
-            className={cn(
-              "flex flex-wrap items-center gap-4 opacity-0",
-              isMounted && playKey ? "animate-hero-cta" : "animate-none"
-            )}
-          >
-            {(() => {
-              const heroCta = resolveCta("無料メルマガ会員登録はこちら");
-              return (
-                <Link
-                  href={heroCta.href}
-                  className="rounded-full bg-win2-primary-orage px-10 py-3 text-sm font-semibold text-white shadow-lg shadow-win2-primary-orage/25 transition hover:bg-win2-accent-amber"
-                >
-                  {heroCta.label}
-                </Link>
-              );
-            })()}
-            <Link
-              href="/login"
-              className="rounded-full border border-win2-accent-rose bg-white px-10 py-3 text-sm font-semibold text-win2-accent-rose transition hover:bg-win2-surface-rose-100"
-            >
-              ログイン
-            </Link>
-          </div>
-        </div>
-        <div
-          ref={imageRef}
-          className={cn(
-            "relative z-20 mx-auto w-full max-w-[420px] opacity-0 md:max-w-[460px]",
-            isMounted && playKey ? "animate-hero-image" : "animate-none"
-          )}
-        >
-          <div className="relative">
-            <Image
-              src="/assets/images/woman.webp"
-              alt="メインビジュアル"
-              width={880}
-              height={880}
-              className="w-full object-contain"
-              priority
-            />
-            {/* 下部を自然にフェードアウト */}
-            <div
-              className="pointer-events-none absolute inset-x-0 bottom-0 h-24 bg-gradient-to-t from-win2-surface-stone-100 to-transparent"
-              style={{ maskImage: 'linear-gradient(to top, black, transparent)' }}
-            />
-          </div>
-        </div>
-      </div>
-      <div className="pointer-events-none absolute inset-x-0 bottom-0 h-28 bg-gradient-to-b from-transparent via-win2-surface-stone-100/80 to-win2-surface-stone-100 md:h-32 md:via-win2-surface-stone-100/90" />
-    </section>
-  );
-}
-
-function ProblemSection() {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="relative z-20 bg-win2-surface-stone-100 pb-16 pt-40 md:pt-48">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[1200px] px-6",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        {/* 見出し */}
-        <div className="mb-12 text-center">
-          <h2 className="flex flex-wrap items-center justify-center gap-2 text-2xl font-bold md:text-3xl lg:text-4xl">
-            <span className="inline-block rounded-full border-2 border-win2-primary-orage bg-white px-4 py-1 text-win2-primary-orage">
-              日常で
-            </span>
-            <span>こんな</span>
-            <span className="text-win2-primary-orage">お悩み</span>
-            <span>、ありませんか？</span>
-          </h2>
-        </div>
-
-        {/* 3カラムレイアウト */}
-        <div className="grid grid-cols-1 items-center gap-8 md:grid-cols-[180px_1fr_180px] lg:grid-cols-[220px_1fr_220px]">
-          {/* 左：悩む人の画像 */}
-          <div className="flex justify-center">
-            <div className="relative h-40 w-40 overflow-hidden rounded-full md:h-44 md:w-44 lg:h-52 lg:w-52">
-              <Image
-                src="/assets/images/problem.webp"
-                alt="悩む人"
-                fill
-                sizes="(max-width: 768px) 160px, (max-width: 1024px) 176px, 208px"
-                className="object-cover grayscale"
-              />
-            </div>
-          </div>
-
-          {/* 中央：お悩みリスト */}
-          <div className="space-y-3 text-left text-sm md:text-base">
-            <p className="leading-relaxed">
-              • 保険料高い気がするけど、
-              <span className="font-semibold text-win2-primary-orage">どこをどう見直せばいいか分からない</span>
-            </p>
-            <p className="leading-relaxed">
-              • 不動産の価値を知りたいけど、
-              <span className="font-semibold text-win2-primary-orage">査定って難しそう</span>
-              で不安
-            </p>
-            <p className="leading-relaxed">
-              • <span className="font-semibold text-win2-primary-orage">転職したい</span>
-              けど、どのサービスが自分に合っているのか分からない
-            </p>
-            <p className="leading-relaxed">
-              • サブスクやキャンペーンが多すぎて、
-              <span className="font-semibold text-win2-primary-orage">どれがお得か比べられない</span>
-            </p>
-            <p className="leading-relaxed">
-              • <span className="font-semibold text-win2-primary-orage">生活費をもっと節約したい</span>
-              けど、何から始めればいいか迷ってる
-            </p>
-          </div>
-
-          {/* 右：女性画像 */}
-          <div className="flex justify-center hidden md:block">
-            <div className="relative h-56 w-40 md:h-64 md:w-44 lg:h-80 lg:w-52">
-              <Image
-                src="/assets/images/woman.webp"
-                alt="案内スタッフ"
-                fill
-                sizes="(max-width: 768px) 160px, (max-width: 1024px) 176px, 208px"
-                className="object-contain"
-              />
-              <div className="pointer-events-none absolute inset-x-0 bottom-0 h-12 bg-gradient-to-b from-transparent to-win2-surface-stone-100" />
-            </div>
-          </div>
-        </div>
-
-        {/* 下部オレンジ帯 */}
-        <div className="relative -mx-6 mt-12 overflow-hidden bg-gradient-to-r from-win2-accent-rose to-win2-accent-amber py-6 text-center md:-mx-0 md:rounded-2xl">
-          <p className="text-lg font-bold text-white md:text-xl">そんな時こそ</p>
-          <p className="mt-1 text-2xl font-bold md:text-3xl">
-            <span className="text-win2-accent-sun">WIN×Ⅱ</span>
-            <span className="text-white">の出番です！</span>
-          </p>
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function IntroductionBoxSection({ resolveCta }: { resolveCta: CtaResolver }) {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="bg-white py-16">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[900px] px-6",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        {/* 吹き出しボックス */}
-        <div className="relative rounded-[40px] border-4 border-win2-primary-orage bg-white p-8 shadow-[0_20px_50px_rgba(242,111,54,0.2)] md:p-10">
-          {/* 上部三角形（オレンジ帯からの矢印効果） */}
-          <div className="absolute -top-6 left-1/2 h-12 w-12 -translate-x-1/2 rotate-45 border-l-4 border-t-4 border-win2-primary-orage bg-white" />
-
-          <div className="relative z-10 space-y-4 text-center">
-            <p className="text-base leading-relaxed md:text-lg">保険・不動産・転職・エンタメなど</p>
-            <p className="text-base leading-relaxed md:text-lg">
-              暮らしに関わる
-              <span className="font-bold text-win2-primary-orage">多彩な情報を1つのサイトでまとめて比較・検討</span>
-            </p>
-            <p className="text-base leading-relaxed md:text-lg">
-              できるから、「知らなかった」で損する前に、
-              <span className="font-bold text-win2-primary-orage">最適な選択肢が見つかります！</span>
-            </p>
-          </div>
-        </div>
-
-        {/* CTAボタン */}
-        <div className="mt-8 text-center">
-          {(() => {
-            const introductionCta = resolveCta("メルマガ会員登録はこちら");
-            return (
-              <Link
-                href={introductionCta.href}
-                className="inline-block rounded-full bg-gradient-to-r from-win2-accent-rose to-win2-primary-orage px-12 py-4 text-base font-bold text-white shadow-lg shadow-win2-accent-rose/30 transition hover:opacity-90 md:text-lg"
-              >
-                {introductionCta.label}
-              </Link>
-            );
-          })()}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function ServiceSection() {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-  const videoRef = useRef<HTMLVideoElement | null>(null);
-  const [hasPlayed, setHasPlayed] = useState(false);
-  const [isVideoError, setIsVideoError] = useState(false);
-
-  useEffect(() => {
-    if (isVideoError) return;
-
-    const videoNode = videoRef.current;
-    if (!videoNode) return;
-
-    if (isVisible && !hasPlayed) {
-      void videoNode
-        .play()
-        .then(() => setHasPlayed(true))
-        .catch(() => {
-          // 自動再生がブロックされた場合はフォールバックとして最後のフレームを表示
-          const fallbackTime = Number.isFinite(videoNode.duration)
-            ? videoNode.duration
-            : 0;
-          videoNode.currentTime = fallbackTime;
-          setHasPlayed(true);
-        });
-    }
-  }, [isVisible, hasPlayed, isVideoError]);
-
-  useEffect(() => {
-    if (isVideoError) return;
-
-    const videoNode = videoRef.current;
-    if (!videoNode) return;
-
-    const handleEnded = () => {
-      videoNode.pause();
-      const endTime = Number.isFinite(videoNode.duration) ? videoNode.duration : 0;
-      videoNode.currentTime = endTime;
-    };
-
-    videoNode.addEventListener("ended", handleEnded);
-    return () => {
-      videoNode.removeEventListener("ended", handleEnded);
-    };
-  }, [isVideoError]);
-
-  return (
-    <section className="bg-win2-surface-cream-50 py-24">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[1080px] px-6 lg:px-8",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        {/* 見出し：WIN×Ⅱは どんなサービス？ */}
-        <div className="mb-12 flex flex-col items-center justify-center gap-6 md:flex-row md:gap-4">
-          <Image
-            src="/assets/images/win2-is-bubble.webp"
-            alt="WIN×Ⅱは"
-            width={180}
-            height={180}
-            className="h-28 w-28 shrink-0 object-contain md:h-32 md:w-32 lg:h-36 lg:w-36"
-            priority
-          />
-          <h2 className="text-3xl font-bold md:text-4xl lg:text-5xl">
-            どんな<span className="text-win2-primary-orage">サービス</span>？
-          </h2>
-        </div>
-
-        {/* カテゴリボタン */}
-        <div className="mb-10 flex flex-wrap justify-center gap-4">
-          {serviceCategories.map((category) => (
-            <div
-              key={category}
-              className="rounded-full bg-white px-6 py-3 text-sm font-semibold text-win2-primary-orage shadow-[0_12px_30px_rgba(242,111,54,0.18)]"
-            >
-              {category}
-            </div>
-          ))}
-        </div>
-        <div className="relative mx-auto w-full max-w-[900px] pt-24 lg:max-w-[1080px]">
-          <div className="overflow-hidden rounded-[32px] bg-win2-surface-cream-50">
-            {isVideoError ? (
-              <Image
-                src="/assets/images/onestop-figure.webp"
-                alt="暮らしを支える4つのカテゴリ"
-                width={720}
-                height={540}
-                className="w-full object-contain"
-              />
-            ) : (
-              <video
-                ref={videoRef}
-                className="block w-full object-cover"
-                playsInline
-                preload="metadata"
-                muted
-                loop={false}
-                controls={false}
-                aria-label="WIN×Ⅱのサービス紹介"
-                poster="/assets/images/onestop-figure.webp"
-                onError={() => setIsVideoError(true)}
-                suppressHydrationWarning
-              >
-                <source src="/assets/images/what-is-win2.webm" type="video/webm" />
-                {/* eslint-disable-next-line @next/next/no-img-element */}
-                <img
-                  src="/assets/images/onestop-figure.webp"
-                  alt="暮らしを支える4つのカテゴリ"
-                  className="block w-full object-contain"
-                />
-              </video>
-            )}
-          </div>
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function MeritSection() {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="bg-white py-24">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[1100px] space-y-12 px-6 text-center lg:px-8",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        <div className="space-y-3">
-          <h2 className="text-3xl font-bold text-win2-accent-amber md:text-4xl">WIN×Ⅱでのメリット</h2>
-          <p className="text-sm text-slate-600 md:text-base">
-            暮らしを変えるヒントが一つに集まる。WIN×Ⅱだからできる 3 つのポイントをご紹介します。
-          </p>
-        </div>
-        <div className="grid gap-8 md:grid-cols-3">
-          {meritItems.map((item) => (
-            <div
-              key={item.number}
-              className="group relative overflow-hidden rounded-[32px] bg-white p-8 shadow-[0_18px_42px_rgba(244,138,60,0.22)] transition-all duration-300 hover:scale-105 hover:shadow-[0_24px_54px_rgba(244,138,60,0.32)]"
-            >
-              {/* 背景グラデーション装飾 */}
-              <div className="absolute -top-20 -right-20 h-40 w-40 rounded-full bg-gradient-to-br from-win2-primary-orage/10 to-win2-accent-amber/10 blur-3xl" />
-
-              {/* POINT バッジ */}
-              <div className="relative mb-6 flex justify-center">
-                <div className="flex h-20 w-20 flex-col items-center justify-center rounded-full bg-gradient-to-br from-win2-primary-orage to-win2-accent-amber text-xs font-bold text-white shadow-lg shadow-win2-primary-orage/30">
-                  <span className="text-[10px] tracking-wider">POINT</span>
-                  <span className="text-2xl">{item.number}</span>
-                </div>
-              </div>
-
-              {/* アイコン画像 */}
-              <div className="relative mb-6 flex justify-center">
-                <div className="relative h-24 w-24">
-                  <Image
-                    src={item.image}
-                    alt={`${item.title}のアイコン`}
-                    fill
-                    sizes="96px"
-                    className="object-contain"
-                  />
-                </div>
-              </div>
-
-              {/* タイトル */}
-              <div className="relative mb-4">
-                <h3 className="text-center text-lg font-bold text-win2-primary-orage md:text-xl">
-                  {item.title}
-                </h3>
-                <div className="mx-auto mt-2 h-1 w-16 rounded-full bg-gradient-to-r from-win2-primary-orage to-win2-accent-amber" />
-              </div>
-
-              {/* 説明文 */}
-              <p className="relative text-center text-sm leading-relaxed text-slate-700 md:text-base">
-                {item.description}
-              </p>
-            </div>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function HighlightSection({ resolveCta, categories }: { resolveCta: CtaResolver; categories: CategoryResponse[] }) {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="relative bg-win2-surface-sky-50 py-24">
-      {/* City background image */}
-      <div className="absolute inset-0">
-        <Image
-          src="/assets/images/city.webp"
-          alt="都市背景"
-          fill
-          sizes="100vw"
-          className="object-cover opacity-30"
-          quality={70}
-        />
-      </div>
-      {/* Gradient overlay for better text readability */}
-      <div className="absolute inset-0 bg-gradient-to-b from-win2-surface-sky-50/80 to-win2-surface-sky-50/90" />
-
-      <div
-        ref={ref}
-        className={cn(
-          "relative space-y-12",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        {/* ヘッダー部分（中央揃え） */}
-        <div className="mx-auto max-w-[1100px] space-y-3 px-6 text-center lg:px-8">
-          <p className="text-sm font-semibold uppercase tracking-[0.35em] text-win2-primary-orage">
-            掲載サービス・活用シーン
-          </p>
-          <h2 className="text-3xl font-bold md:text-4xl">暮らしを変える多彩なサービス</h2>
-          <p className="text-sm text-slate-600 md:text-base">
-            家計のお悩みからライフイベントまで、WIN×Ⅱなら幅広いカテゴリをワンストップでチェックできます。
-          </p>
-        </div>
-
-        {/* スクロールセクション（横幅いっぱい） */}
-        <div className="relative">
-          {/* グラデーションフェード（左） */}
-          <div className="pointer-events-none absolute inset-y-0 left-0 z-10 w-20 bg-gradient-to-r from-win2-surface-sky-50 via-win2-surface-sky-50/80 to-transparent md:w-32" />
-          {/* グラデーションフェード（右） */}
-          <div className="pointer-events-none absolute inset-y-0 right-0 z-10 w-20 bg-gradient-to-l from-win2-surface-sky-50 via-win2-surface-sky-50/80 to-transparent md:w-32" />
-
-          <div
-            className="hide-scrollbar relative flex snap-x snap-mandatory gap-6 overflow-x-auto px-6 pb-6 pt-1 md:gap-8 md:px-8 lg:px-12"
-            aria-roledescription="horizontal carousel"
-          >
-            {categories.map((category, index) => (
-              <Link
-                key={category.id}
-                href={`/category/${category.id}`}
-                aria-label={`${category.name}の特集を見る`}
-                className={cn(
-                  "group relative isolate flex w-[85vw] min-h-[420px] min-w-[280px] max-w-[360px] flex-col justify-between overflow-hidden rounded-2xl bg-slate-900 text-left text-white shadow-md transition-all duration-300 ease-out",
-                  "focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-win2-accent-amber/50",
-                  "sm:w-[420px] sm:min-w-[380px] md:w-[460px] lg:w-[500px] lg:min-h-[560px]",
-                  "snap-start hover:shadow-lg focus-visible:shadow-lg"
-                )}
-              >
-                <Image
-                  src={category.image?.url || "/assets/images/city.webp"}
-                  alt={category.image?.alt || category.name}
-                  fill
-                  sizes="(min-width: 1024px) 30vw, 85vw"
-                  className="absolute inset-0 h-full w-full object-cover transition-transform duration-500 ease-out group-hover:scale-105"
-                  priority={index === 0}
-                />
-                <span className="absolute inset-0 bg-gradient-to-b from-black/40 to-black/60" />
-
-                <div className="relative flex flex-col gap-4 p-6 sm:p-8">
-                  <div className="space-y-3">
-                    <span className="inline-flex items-center rounded-md bg-white/10 px-3 py-1 text-xs font-medium text-white/90">
-                      {category.name}
-                    </span>
-                    <h3 className="text-2xl font-bold leading-tight sm:text-3xl">{category.name}</h3>
-                    <p className="text-sm leading-relaxed text-white/80 sm:text-base">
-                      {category.description || "カテゴリの詳細をご覧ください"}
-                    </p>
-                  </div>
-                </div>
-              </Link>
-            ))}
-          </div>
-        </div>
-
-        {/* CTA ボタン（中央揃え） */}
-        <div className="mx-auto max-w-[1100px] px-6 text-center lg:px-8">
-          {(() => {
-            const highlightCta = resolveCta("無料メルマガ会員登録で最新情報を受け取る");
-            return (
-              <Link
-                href={highlightCta.href}
-                className="inline-flex items-center justify-center rounded-full bg-gradient-to-r from-win2-accent-rose to-win2-primary-orage px-12 py-3 text-base font-semibold text-white shadow-lg shadow-win2-accent-rose/30 transition hover:opacity-90 focus-visible:outline-none focus-visible:ring-4 focus-visible:ring-win2-accent-rose/40"
-              >
-                {highlightCta.label}
-              </Link>
-            );
-          })()}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function AchievementSection() {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="bg-win2-surface-cream-50 py-24">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[960px] space-y-12 px-6 text-center lg:px-8",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        <h2 className="text-3xl font-bold text-win2-accent-amber md:text-4xl">実績・掲載数</h2>
-        <div className="grid gap-8 md:grid-cols-2">
-          {achievementImages.map((achievement) => (
-            <div
-              key={achievement.alt}
-              className="rounded-[32px] bg-white px-6 py-10 shadow-[0_20px_45px_rgba(244,138,60,0.2)]"
-            >
-              <Image
-                src={achievement.image}
-                alt={achievement.alt}
-                width={420}
-                height={320}
-                className="mx-auto w-full max-w-[320px] object-contain"
-              />
-            </div>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function TestimonialsSection() {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="bg-white py-24">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[1080px] space-y-10 px-6 text-center lg:px-8",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        <div className="space-y-2">
-          <h2 className="text-3xl font-bold md:text-4xl">ご利用者様の声</h2>
-          <p className="text-sm text-slate-600 md:text-base">
-            WIN×Ⅱをご利用いただいた皆さまからのお声を一部ご紹介します。
-          </p>
-        </div>
-        <div className="grid gap-6 md:grid-cols-2">
-          {testimonials.map((testimonial, index) => (
-            <div
-              key={`${testimonial.age}-${testimonial.gender}-${index}`}
-              className="group relative rounded-[24px] bg-gradient-to-br from-win2-surface-cream-320 to-white p-6 text-left shadow-[0_12px_28px_rgba(0,0,0,0.06)] transition-all duration-300 hover:shadow-[0_20px_40px_rgba(0,0,0,0.12)]"
-            >
-              {/* 引用符デザイン */}
-              <div className="absolute top-4 left-4 text-6xl font-serif text-win2-primary-orage/20">&ldquo;</div>
-
-              {/* ユーザー情報 */}
-              <div className="relative mb-4 flex items-center gap-3">
-                {/* アバター画像 */}
-                <div className="relative h-14 w-14 shrink-0 overflow-hidden rounded-full shadow-md ring-2 ring-win2-primary-orage/20">
-                  <Image
-                    src={testimonial.avatar}
-                    alt={`${testimonial.age} ${testimonial.gender}`}
-                    fill
-                    sizes="56px"
-                    className="object-cover"
-                  />
-                </div>
-                <div>
-                  {/* 年齢・性別バッジ */}
-                  <div className="inline-block rounded-full bg-gradient-to-r from-win2-accent-rose to-win2-accent-amber px-3 py-1 text-xs font-semibold text-white">
-                    {testimonial.age} {testimonial.gender}
-                  </div>
-                  {/* 星評価 */}
-                  <div className="mt-1 text-sm">
-                    {Array.from({ length: testimonial.rating }).map((_, i) => (
-                      <span key={i} className="text-[#ffd700]">⭐</span>
-                    ))}
-                  </div>
-                </div>
-              </div>
-
-              {/* コメント */}
-              <p className="relative z-10 text-sm leading-relaxed text-slate-700 md:text-base">
-                {testimonial.comment}
-              </p>
-            </div>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function FaqSection() {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-  const [openIndex, setOpenIndex] = useState<number | null>(null);
-
-  const toggleFaq = (index: number) => {
-    setOpenIndex(openIndex === index ? null : index);
-  };
-
-  return (
-    <section className="bg-win2-surface-cream-300 py-24">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto max-w-[960px] space-y-10 px-6 text-center lg:px-8",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        <div className="space-y-2">
-          <h2 className="text-3xl font-bold md:text-4xl">よくあるご質問</h2>
-          <p className="text-sm text-slate-600 md:text-base">
-            ご不明な点があれば、お気軽にお問い合わせください。
-          </p>
-        </div>
-        <div className="space-y-4 text-left">
-          {faqItems.map((item, index) => {
-            const isOpen = openIndex === index;
-            return (
-              <div
-                key={item.question}
-                className="overflow-hidden rounded-[24px] bg-white shadow-[0_12px_26px_rgba(0,0,0,0.08)] transition-all duration-300 hover:shadow-[0_18px_36px_rgba(0,0,0,0.14)]"
-              >
-                {/* 質問部分（クリック可能） */}
-                <button
-                  onClick={() => toggleFaq(index)}
-                  className="flex w-full items-start gap-3 p-6 text-left transition-colors hover:bg-win2-surface-cream-300"
-                >
-                  {/* Qバッジ */}
-                  <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-win2-primary-orage to-win2-accent-amber text-base font-bold text-white shadow-md">
-                    Q
-                  </div>
-                  <div className="flex-1">
-                    <p className="text-sm font-semibold text-win2-primary-orage md:text-base">
-                      {item.question}
-                    </p>
-                  </div>
-                  {/* 開閉アイコン */}
-                  <div
-                    className={cn(
-                      "shrink-0 text-win2-primary-orage transition-transform duration-300",
-                      isOpen ? "rotate-180" : ""
-                    )}
-                  >
-                    <svg
-                      className="h-5 w-5"
-                      fill="none"
-                      stroke="currentColor"
-                      viewBox="0 0 24 24"
-                    >
-                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 9l-7 7-7-7" />
-                    </svg>
-                  </div>
-                </button>
-
-                {/* 回答部分（アコーディオン） */}
-                <div
-                  className={cn(
-                    "overflow-hidden transition-all duration-300",
-                    isOpen ? "max-h-96 opacity-100" : "max-h-0 opacity-0"
-                  )}
-                >
-                  <div className="border-t border-slate-200 bg-gradient-to-br from-[#fafafa] to-white px-6 pb-6 pt-4">
-                    <div className="flex items-start gap-3">
-                      {/* Aバッジ */}
-                      <div className="flex h-8 w-8 shrink-0 items-center justify-center rounded-full bg-gradient-to-br from-[#3b82f6] to-[#2563eb] text-sm font-bold text-white shadow-md">
-                        A
-                      </div>
-                      <p className="flex-1 text-sm leading-relaxed text-slate-700 md:text-base">
-                        {item.answer}
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            );
-          })}
-        </div>
-      </div>
-    </section>
-  );
-}
-
-function BottomCtaSection({ resolveCta }: { resolveCta: CtaResolver }) {
-  const { ref, isVisible } = useScrollReveal<HTMLDivElement>();
-
-  return (
-    <section className="bg-gradient-to-r from-win2-accent-rose to-win2-accent-amber py-16">
-      <div
-        ref={ref}
-        className={cn(
-          "mx-auto flex max-w-[860px] flex-col items-center gap-5 px-6 text-center text-white lg:px-8",
-          "transition-transform-opacity",
-          isVisible ? "reveal-visible" : "reveal"
-        )}
-      >
-        <h2 className="text-3xl font-bold md:text-4xl">いますぐ WIN×Ⅱ をはじめましょう</h2>
-        <p className="text-sm leading-relaxed text-white/90 md:text-base">
-          暮らしをもっとお得に、もっとスマートに。WIN×Ⅱの無料メルマガ会員登録でキャッシュバック特典と最新情報を手に入れてください。
-        </p>
-        <div className="flex flex-wrap justify-center gap-4">
-          {(() => {
-            const bottomCta = resolveCta("無料メルマガ会員登録はこちら");
-            return (
-              <Link
-                href={bottomCta.href}
-                className="rounded-full bg-white px-10 py-3 text-sm font-semibold text-win2-accent-rose transition hover:bg-white/90"
-              >
-                {bottomCta.label}
-              </Link>
-            );
-          })()}
-          <Link
-            href="/blog"
-            className="rounded-full border border-white/80 px-10 py-3 text-sm font-semibold text-white transition hover:bg-white/10"
-          >
-            ブログで最新情報を見る
-          </Link>
-        </div>
-      </div>
-    </section>
   );
 }
