+++ components/home/types.ts
@@ -1,0 +1,1 @@
+export type CtaResolver = (defaultLabel: string) => { href: string; label: string };
